Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -6,7 +6,7 @@ int main() {
     scanf("%d\n",&a);
     scanf("%d\n",&b);
     scanf("%c",&ch);
-    div=(float)a/b;
+    div=a/b;
     if(ch=='+'){
         printf("%d",a+b);
     } else if(ch=='-'){
